+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 6.0.0-dev.240 (2026-05-11)
+
+### 🚀 Features
+
+- gigantic hub and spoke catchup integration test ([4283fead6](https://github.com/powerhouse-inc/powerhouse/commit/4283fead6))
+
+### 🩹 Fixes
+
+- **reactor-api:** exclude hub/spoke test by default, added specific job to test it ([8e8474929](https://github.com/powerhouse-inc/powerhouse/commit/8e8474929))
+- **release:** pass the just-published tag from release -> publish-ph-binaries ([dd19a9b20](https://github.com/powerhouse-inc/powerhouse/commit/dd19a9b20))
+
+### ❤️ Thank You
+
+- acaldas
+- Benjamin Jordan
+
 ## 6.0.0-dev.239 (2026-05-11)
 
 ### 🚀 Features